Add support for faster uploading via UART
When mox-imager is used to send boot image via UART (-D option), it now
accepts -b BAUD / --baudrate=BAUD option.

When used, the TIM images are send via standard baudrate 115200, but
then a baudrate change request is sent.

This has to be supported on the other side via a GPP package
interpreting the baudrate change request.

Signed-off-by: Marek Behún <marek.behun@nic.cz>