If OBMI image / a53-firmware is given when deploying, send it
This can be useful in a scenario where we want to deploy a trusted
device in one go - writing OTP rows for trusted boot and then flashing
boot device with a trusted firmware image. The WTMI firmware we are
sending when deploying will write OTPs and then boot the OBMI image,
which can flash boot device.

Signed-off-by: Marek Behún <kabel@kernel.org>